Add: A new Schema Add: Resolvers Add: removing the mocks
