+++ src/index.js
@@ -1,12 +1,10 @@
 const { ApolloServer } = require('apollo-server');
 const typeDefs = require('./schema');
-const mocks = require('./mocks');
-// const resolvers = require('./resolvers');
+const resolvers = require('./resolvers');
 
 const server = new ApolloServer({
   typeDefs,
-  mocks
-  // resolvers
+  resolvers
 });
 
 server.listen().then(() => {
